renamed compare interfaces with matcher and made it more flexible for any type of interceptions
