+++ src/platform/http/interceptors/http-interceptor-mapping.interface.ts
@@ -1,0 +1,6 @@
+import { IHttpInterceptor } from './http-interceptor.interface';
+
+export interface IHttpInterceptorMapping {
+  interceptor: IHttpInterceptor;
+  paths: string[];
+}
