+++ src/platform/http/interceptors/http-interceptor-matcher.interface.ts
@@ -1,0 +1,9 @@
+import { RequestOptionsArgs } from '@angular/http';
+
+import { IHttpInterceptorMapping } from './http-interceptor-mapping.interface';
+
+export interface IHttpInterceptorMatcher {
+
+  matches(options: RequestOptionsArgs, mapping: IHttpInterceptorMapping): boolean;
+
+}
