--- src/platform/http/interceptors/url-compare.interface.ts
@@ -1,5 +1,0 @@
-export interface IURLCompare {
-
-  compare(url: string, paths: string[]): boolean;
-
-}
